Fixed and optimized the code

+++ Bagger/BossHelper.cs
@@ -12,28 +12,28 @@ public static class BossHelper
     /// </summary>
     public static readonly Dictionary<int, int> BossMasks = new()
     {
-        { NPCID.KingSlime, 1 << 0 },           // 1
-        { NPCID.EyeofCthulhu, 1 << 1 },        // 2
-        { NPCID.EaterofWorldsHead, 1 << 2 },  // 4
-        { NPCID.EaterofWorldsBody, 1 << 2 },  // 4
-        { NPCID.EaterofWorldsTail, 1 << 2 },  // 4
-        { NPCID.BrainofCthulhu, 1 << 3 },     // 8
-        { NPCID.QueenBee, 1 << 4 },           // 16
-        { NPCID.SkeletronHead, 1 << 5 },      // 32
-        { NPCID.Deerclops, 1 << 6 },          // 64
-        { NPCID.WallofFlesh, 1 << 7 },        // 128
-        { NPCID.QueenSlimeBoss, 1 << 8 },     // 256
-        { NPCID.TheDestroyer, 1 << 9 },       // 512
-        { NPCID.Retinazer, 1 << 10 },         // 1024
-        { NPCID.Spazmatism, 1 << 10 },        // 1024
-        { NPCID.SkeletronPrime, 1 << 11 },    // 2048
-        { NPCID.Plantera, 1 << 12 },          // 4096
-        { NPCID.Golem, 1 << 13 },             // 8192
-        { NPCID.DukeFishron, 1 << 14 },       // 16384
-        { NPCID.HallowBoss, 1 << 15 },        // 32768
-        { NPCID.CultistBoss, 1 << 16 },       // 65536
-        { NPCID.DD2Betsy, 1 << 17 },          // 131072
-        { NPCID.MoonLordCore, 1 << 18 }       // 262144
+        { NPCID.KingSlime, 1 << 0 },
+        { NPCID.EyeofCthulhu, 1 << 1 },
+        { NPCID.EaterofWorldsHead, 1 << 2 },
+        { NPCID.EaterofWorldsBody, 1 << 2 },
+        { NPCID.EaterofWorldsTail, 1 << 2 },
+        { NPCID.BrainofCthulhu, 1 << 3 },
+        { NPCID.QueenBee, 1 << 4 },
+        { NPCID.SkeletronHead, 1 << 5 },
+        { NPCID.Deerclops, 1 << 6 },
+        { NPCID.WallofFlesh, 1 << 7 },
+        { NPCID.QueenSlimeBoss, 1 << 8 },
+        { NPCID.TheDestroyer, 1 << 9 },
+        { NPCID.Retinazer, 1 << 10 },
+        { NPCID.Spazmatism, 1 << 10 },
+        { NPCID.SkeletronPrime, 1 << 11 },
+        { NPCID.Plantera, 1 << 12 },
+        { NPCID.Golem, 1 << 13 },
+        { NPCID.DukeFishron, 1 << 14 },
+        { NPCID.HallowBoss, 1 << 15 },
+        { NPCID.CultistBoss, 1 << 16 },
+        { NPCID.DD2Betsy, 1 << 17 },
+        { NPCID.MoonLordCore, 1 << 18 }
     };
 
     /// <summary>
@@ -41,25 +41,25 @@ public static class BossHelper
     /// </summary>
     public static readonly Dictionary<string, int[]> ConfigKeyToBossIds = new()
     {
-        { "KingSlime", new[] { NPCID.KingSlime } },
-        { "EyeOfCthulhu", new[] { NPCID.EyeofCthulhu } },
-        { "EaterOfWorlds", new[] { NPCID.EaterofWorldsHead } },
-        { "BrainOfCthulhu", new[] { NPCID.BrainofCthulhu } },
-        { "QueenBee", new[] { NPCID.QueenBee } },
-        { "Skeletron", new[] { NPCID.SkeletronHead } },
-        { "Deerclops", new[] { NPCID.Deerclops } },
-        { "WallOfFlesh", new[] { NPCID.WallofFlesh } },
-        { "QueenSlime", new[] { NPCID.QueenSlimeBoss } },
-        { "TheDestroyer", new[] { NPCID.TheDestroyer } },
-        { "TheTwins", new[] { NPCID.Retinazer, NPCID.Spazmatism } },
-        { "SkeletronPrime", new[] { NPCID.SkeletronPrime } },
-        { "Plantera", new[] { NPCID.Plantera } },
-        { "Golem", new[] { NPCID.Golem } },
-        { "DukeFishron", new[] { NPCID.DukeFishron } },
-        { "EmpressOfLight", new[] { NPCID.HallowBoss } },
-        { "LunaticCultist", new[] { NPCID.CultistBoss } },
-        { "Betsy", new[] { NPCID.DD2Betsy } },
-        { "MoonLord", new[] { NPCID.MoonLordCore } }
+        { "KingSlime", new[] { (int)NPCID.KingSlime } },
+        { "EyeOfCthulhu", new[] { (int)NPCID.EyeofCthulhu } },
+        { "EaterOfWorlds", new[] { (int)NPCID.EaterofWorldsHead } },
+        { "BrainOfCthulhu", new[] { (int)NPCID.BrainofCthulhu } },
+        { "QueenBee", new[] { (int)NPCID.QueenBee } },
+        { "Skeletron", new[] { (int)NPCID.SkeletronHead } },
+        { "Deerclops", new[] { (int)NPCID.Deerclops } },
+        { "WallOfFlesh", new[] { (int)NPCID.WallofFlesh } },
+        { "QueenSlime", new[] { (int)NPCID.QueenSlimeBoss } },
+        { "TheDestroyer", new[] { (int)NPCID.TheDestroyer } },
+        { "TheTwins", new[] { (int)NPCID.Retinazer, (int)NPCID.Spazmatism } },
+        { "SkeletronPrime", new[] { (int)NPCID.SkeletronPrime } },
+        { "Plantera", new[] { (int)NPCID.Plantera } },
+        { "Golem", new[] { (int)NPCID.Golem } },
+        { "DukeFishron", new[] { (int)NPCID.DukeFishron } },
+        { "EmpressOfLight", new[] { (int)NPCID.HallowBoss } },
+        { "LunaticCultist", new[] { (int)NPCID.CultistBoss } },
+        { "Betsy", new[] { (int)NPCID.DD2Betsy } },
+        { "MoonLord", new[] { (int)NPCID.MoonLordCore } }
     };
 
     public static int AddBossToMask(int mask, int npcType)
